+++ packages/core/src/index.tsx
@@ -1,9 +1,3 @@
-// core
 export { bind } from "./bind"
-export { shareLatest } from "./operators/shareLatest"
-
-// support for React Suspense
+export { shareLatest } from "./shareLatest"
 export { SUSPENSE } from "./SUSPENSE"
-export { suspend } from "./operators/suspend"
-export { suspended } from "./operators/suspended"
-export { switchMapSuspended } from "./operators/switchMapSuspended"
