+++ packages/utils/src/index.tsx
@@ -4,3 +4,6 @@ export { collectValues } from "./collectValues"
 export { collect } from "./collect"
 export { mergeWithKey } from "./mergeWithKey"
 export { split } from "./split"
+export { suspend } from "./suspend"
+export { suspended } from "./suspended"
+export { switchMapSuspended } from "./switchMapSuspended"
