+++ packages/utils/src/suspend.ts
@@ -1,6 +1,6 @@
 import { ObservableInput, from } from "rxjs"
 import { startWith } from "rxjs/operators"
-import { SUSPENSE } from "../SUSPENSE"
+import { SUSPENSE } from "@react-rxjs/core"
 
 /**
  * A RxJS creation operator that prepends a SUSPENSE on the source observable.
